(PC-19593) refactor(navigation): move useInitialScreenConfig tests outside __tests__ folder
